requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
fitz
pymupdf
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain_openai
beautifulsoup4
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse-starlette
openai
langchain_ollama